style : 마이페이지 컴포넌트 분리 스타일작업
