+++ src/components/mypage/ProfileHeader.tsx
@@ -13,18 +13,18 @@ const ProfileHeader = () => {
   const { postCount, likedPostCount, metPeopleCount } = profileCount();
 
   const buttonData = [
-    { icon: GoPeople, title: '스쳐간 인연', count: metPeopleCount },
-    { icon: BsPostcardHeart, title: '작성 글', count: postCount },
-    { icon: IoMdHeartEmpty, title: '좋아요한 글', count: likedPostCount }
+    { icon: GoPeople, title: '스쳐간 인연 >', count: metPeopleCount },
+    { icon: BsPostcardHeart, title: '내가 쓴 글 >', count: postCount },
+    { icon: IoMdHeartEmpty, title: '좋아요한 글 >', count: likedPostCount }
   ];
 
   return (
-    <div className="bg-purpleSecondary w-full py-[40px] px-[24px]">
-      <div className="flex flex-col gap-4 max-w-[1000px] m-auto max-md:items-start">
-        <h1 className="text-[42px] font-bold">프로필</h1>
+    <div className="bg-mainColor w-full py-[48px] px-[24px] text-white">
+      <div className="flex flex-col gap-6 max-w-[1280px] m-auto max-md:items-start">
+        <h1 className="text-2xl font-medium">마이페이지</h1>
         <div className="flex items-center max-md:items-start md:justify-between max-md:flex-col max-md:gap-6">
           <div className="flex gap-6 items-center">
-            <div className="max-md:w-[160px] max-md:h-[160px] w-[180px] h-[180px] overflow-hidden flex justify-center items-center rounded-full relative">
+            <div className="w-[108px] h-[108px] overflow-hidden flex justify-center items-center rounded-full relative">
               {user?.avatar ? (
                 <Image
                   src={`${user?.avatar}?${new Date().getTime()}`}
@@ -35,18 +35,34 @@ const ProfileHeader = () => {
                   priority={true}
                 />
               ) : (
-                <Avatar color="secondary" className="w-32 h-32" />
+                <Avatar color="default" className="w-32 h-32" />
               )}
             </div>
             <div className="flex flex-col gap-2">
-              <p className="block text-2xl font-semibold">{user?.nickname}</p>
-              <p className="block text-lg font-medium">{user?.login_email}</p>
-              <p className="font-medium text-gray3 text-sm">가입일 : {joinTime}</p>
+              <div className="flex items-center max-lg:flex-col max-lg:items-start gap-2">
+                <p className="block text-2xl font-bold">{user?.nickname}</p>
+                <p className="text-2xl font-thin max-lg:hidden">|</p>
+                <div>
+                  <span className="text-sm tracking-wider">카카오톡 ID : </span>
+                  <span className="text-sm font-medium tracking-wider">{user?.kakaoId}</span>
+                </div>
+              </div>
+              <div className="flex items-center max-lg:flex-col max-lg:items-start gap-2">
+                {user?.isValidate ? (
+                  <p className="px-[12px] py-[4px] bg-white rounded-full text-warningYellow text-sm font-semibold">
+                    대학교 인증회원
+                  </p>
+                ) : (
+                  <p></p>
+                )}
+                <p className="block text-sm tracking-wider">{user?.school_email}</p>
+              </div>
+              <p className="font-light text-sm">가입일 : {joinTime}</p>
             </div>
           </div>
-          <div className="flex gap-10 max-md:gap-6">
+          <div className="flex">
             {buttonData.map((item, index) => (
-              <button key={index} className="font-semibold flex flex-col items-center gap-4 max-md:gap-2">
+              <button key={index} className="flex flex-col items-start gap-2 w-40 max-lg:w-32">
                 <item.icon size={40} />
                 <p>{item.title}</p>
                 <p className="font-bold text-3xl">{item.count}</p>
